Isolate the DA Application class
ISolate the DA Application class from the rest of the implementation so that we can handle setup an Assembly resolver before all the dependencies need to get solved.
